finish 1.1

+++ src/storage/disk_manager.cpp
@@ -14,21 +14,24 @@ DiskManager::DiskManager() { memset(fd2pageno_, 0, MAX_FD * (sizeof(std::atomic<
  *
  */
 void DiskManager::write_page(int fd, page_id_t page_no, const char *offset, int num_bytes) {
-    // TODO:
+    // todo:
     //  1.lseek()定位到文件头，通过(fd,page_no)可以定位指定页面及其在磁盘文件中的偏移量
     //  2.调用write()函数
     //  注意处理异常
-    // lseek(fd,0,SEEK_SET);
+    lseek(fd,page_no*PAGE_SIZE,SEEK_SET);
+    write(fd,offset,num_bytes);
 }
 
 /**
  * @brief Read the contents of the specified page into the given memory area
  */
 void DiskManager::read_page(int fd, page_id_t page_no, char *offset, int num_bytes) {
-    // TODO:
+    // todo:
     //  1.lseek()定位到文件头，通过(fd,page_no)可以定位指定页面及其在磁盘文件中的偏移量
     //  2.调用read()函数
     //  注意处理异常
+    lseek(fd,page_no*PAGE_SIZE,SEEK_SET);
+    read(fd,offset,num_bytes);
 }
 
 /**
@@ -36,7 +39,7 @@ void DiskManager::read_page(int fd, page_id_t page_no, char *offset, int num_byt
  * For now just keep an increasing counter
  */
 page_id_t DiskManager::AllocatePage(int fd) {
-    // TODO:
+    // todo:
     //  简单的自增分配策略，指定文件的页面编号加1
     set_fd2pageno(fd,get_fd2pageno(fd)+1);
     return -1;
@@ -73,23 +76,30 @@ void DiskManager::destroy_dir(const std::string &path) {
  * @brief 用于判断指定路径文件是否存在
  */
 bool DiskManager::is_file(const std::string &path) {
-    // TODO:FINISH
+    // todo:FINISH
     //  用struct stat获取文件信息
     //
     struct stat st;
-    return stat(path.c_str(), &st) == 0;
+    return stat(path.c_str(), &st) == 0 && S_ISREG(st.st_mode);
 }
 
 /**
  * @brief 用于创建指定路径文件
  */
 void DiskManager::create_file(const std::string &path) {
-    // TODO:FINISH
+    // todo:FINISH
     //  调用open()函数，使用O_CREAT模式
     //  注意不能重复创建相同文件
-    if (is_file(path)) return;
-    int fd = open(path.c_str(), O_CREAT);
-    if (fd == -1) return;
+    if (is_file(path))
+    {
+        throw FileExistsError(path);
+        return;
+    }
+    int fd = open(path.c_str(), O_CREAT|O_RDWR,0600);
+    if (fd == -1)
+    {
+        
+    }
     return;
 }
 
@@ -97,9 +107,15 @@ void DiskManager::create_file(const std::string &path) {
  * @brief 用于删除指定路径文件
  */
 void DiskManager::destroy_file(const std::string &path) {
-    // TODO:FINISH
+    // todo:FINISH
     //  调用unlink()函数
     //  注意不能删除未关闭的文件
+    if(!is_file(path))
+    {
+        throw FileNotFoundError(path);
+        return;
+    }
+    // 说明打开文件列表里面没有
     if (path2fd_.find(path) == path2fd_.end()) {
         unlink(path.c_str());
     }
@@ -109,9 +125,14 @@ void DiskManager::destroy_file(const std::string &path) {
  * @brief 用于打开指定路径文件
  */
 int DiskManager::open_file(const std::string &path) {
-    // TODO:FINISH
+    // todo:FINISH
     //  调用open()函数，使用O_RDWR模式
     //  注意不能重复打开相同文件，并且需要更新文件打开列表
+    if(!is_file(path))
+    {
+        throw FileNotFoundError(path);
+    }
+    // 说明打开文件列表里面没有
     if (path2fd_.find(path) == path2fd_.end()) {
         int fd = open(path.c_str(), O_RDWR);
         path2fd_.insert({path, fd});
@@ -125,12 +146,18 @@ int DiskManager::open_file(const std::string &path) {
  * @brief 用于关闭指定路径文件
  */
 void DiskManager::close_file(int fd) {
-    // TODO:FINISH
+    // todo:FINISH
     //  调用close()函数
     //  注意不能关闭未打开的文件，并且需要更新文件打开列表
-    if(fd2path_.find(fd)==fd2path_.end())
+    
+    // 说明打开文件列表里面有
+    if(fd2path_.find(fd)!=fd2path_.end())
     {
-        std::string path = fd2path_[fd];
+        std::string path = fd2path_.at(fd);
+        if(!is_file(path))
+        {
+            throw FileNotFoundError(path);
+        }
         close(fd);
         path2fd_.erase(path);
         fd2path_.erase(fd);
